+++ packages/ListBox/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed hardcoded text in listbox footer.
+
 ## [1.0.0] - 2020-11-04
 
 ### Changed